requirements.txt
gunicorn==20.1.0
Flask==2.0.3
tensorflow==2.12.0
Pillow==9.5.0
numpy==1.23.5